refactor(SideBarMenu): Optimize app icons usage
This commit changes the way app icons are imported and used within the
SideMenu component. Icons are now exported as constants from their
respective files, and directly used in the SideMenu.

This change enhances performance by avoiding the creation of new
instances for each icon every time the SideMenu is rendered.