feat(i18n): Implement i18n support and add EN and PT-BR translations
This commit includes the implementation of i18next for
internationalization (i18n) support in the project.

All text in the application has been translated into English
and Brazilian Portuguese.

Namespaces are used to organize translations.

A custom hook is used to ensure reactivity of menu items when
the language is changed.